[ROC-74]  Implementacja przydzielania zawodników do meczy (#30)
* update user_id from long to integer to match db schema

* implement first draft of starting the competition

* update slots assignment, add limiting participants

* update limit calculation, add test

* add TimeSlots class

* do not allow registration for closed competitions

* after review changes

* prepare for frontend requests

* frontend update

* add missing data fetching

* change buttons colors

+++ backend/src/main/java/com/roc/app/competition/assignment/CompetitionParticipant.java
@@ -21,7 +21,7 @@ public class CompetitionParticipant {
 
     @Id
     @Column(name = FieldNames.PARTICIPANT_ID)
-    private Long participantId;
+    private Integer participantId;
 
     @Enumerated(EnumType.STRING)
     @Column(name = FieldNames.PARTICIPANT_STATUS, nullable = false)
@@ -35,7 +35,7 @@ public class CompetitionParticipant {
     @AllArgsConstructor
     public static class CompetitionParticipantId implements Serializable {
         private Integer competitionId;
-        private Long participantId;
+        private Integer participantId;
     }
 
     public static final class FieldNames {
